--- pkg/network/java/testdata/.gitignore
@@ -1,1 +1,0 @@
-TestAgentLoaded.agentmain.*
