+++ pkg/network/protocols/tls/java/testdata/.gitignore
@@ -1,0 +1,1 @@
+TestAgentLoaded.agentmain.130289700
