+++ pkg/network/protocols/tls/java/testdata/docker-compose.yml
@@ -1,0 +1,16 @@
+version: '3'
+services:
+  java:
+    image: ${IMAGE_VERSION}
+    entrypoint: java -cp /v ${ENTRYCLASS}
+    extra_hosts:
+      - ${EXTRA_HOSTS}
+    # setting the limits here as new os confuse (abort()) old glibc with default big numbers
+    # https://datadoghq.atlassian.net/browse/USMO-295
+    # numbers here are default from /etc/systemd/system.conf
+    ulimits:
+      nofile:
+        soft: 1024
+        hard: 524288
+    volumes:
+      - ./:/v:z
